initial refactor of enriched document into speech document
