+++ components/mjs/a11y/speech/config.json
@@ -1,0 +1,15 @@
+{
+  "build": {
+    "component": "a11y/speech",
+    "targets": ["a11y/speech.ts", "a11y/speech"]
+  },
+  "webpack": {
+    "name": "a11y/speech",
+    "libs": [
+      "components/src/a11y/semantic-enrich/lib",
+      "components/src/a11y/sre/lib",
+      "components/src/input/mml/lib",
+      "components/src/core/lib"
+    ]
+  }
+}
